+++ tests/testthat/test-pars.R
@@ -14,11 +14,28 @@ test_that("pars.term", {
   ))
   expect_identical(pars(terms), c("alpha", "beta", "sigma"))
   expect_identical(pars(terms, scalar = TRUE), "sigma")
-  expect_identical(pars(terms, scalar = TRUE, terms = TRUE), "sigma")
-  expect_identical(
-    pars(terms, terms = TRUE),
-    c("alpha", "alpha", "beta", "beta", "beta", "beta", "sigma")
-  )
+})
+
+test_that("pars.term deprecated terms", {
+  terms <- as.term(c(
+    "alpha[1]", "alpha[2]", "beta[1,1]", "beta[2,1]",
+    "beta[1,2]", "beta[2,2]", "sigma"
+  ))
+  
+  rlang::with_options(lifecycle_verbosity = "error", {
+    expect_error(pars(terms, scalar = TRUE, terms = TRUE), class = "defunctError")
+  })
+  
+  rlang::with_options(lifecycle_verbosity = "quiet", {
+    expect_identical(pars(terms, scalar = TRUE, terms = TRUE), "sigma")
+  })
+  
+  rlang::with_options(lifecycle_verbosity = "quiet", {
+    expect_identical(
+      pars(terms, terms = TRUE),
+      c("alpha", "alpha", "beta", "beta", "beta", "beta", "sigma")
+    )
+  })  
 })
 
 test_that("pars.term", {
@@ -42,21 +59,7 @@ test_that("pars.term missing values", {
     pars(as.term(c(NA_character_, "a", NA_character_, "a"))),
     c(NA_character_, "a")
   )
-
-  expect_identical(pars(as.term(NA_character_), term = TRUE), NA_character_)
-  expect_identical(
-    pars(as.term(c(NA_character_, "a")), term = TRUE),
-    c(NA_character_, "a")
-  )
-  expect_identical(
-    pars(as.term(c("a", NA_character_, "a")), term = TRUE),
-    c("a", NA_character_, "a")
-  )
-  expect_identical(
-    pars(as.term(c(NA_character_, "a", NA_character_, "a")), term = TRUE),
-    c(NA_character_, "a", NA_character_, "a")
-  )
-
+  
   expect_identical(pars(as.term(NA_character_), scalar = TRUE), NA_character_)
   expect_identical(
     pars(as.term(c(NA_character_, "a")), scalar = TRUE),
@@ -102,7 +105,7 @@ test_that("set_pars", {
     set_pars(as.term("a"), "1"),
     "^`value` must be a valid parameter name[.]$"
   )
-
+  
   expect_identical(set_pars(as.term(c("a", "b")), c("b", "a")), as.term(c("b", "a")))
   expect_identical(set_pars(as.term(c("a", "b")), c("b", "d")), as.term(c("b", "d")))
   expect_identical(set_pars(as.term(c("a [ 1]", "b")), c("b", "d")), as.term(c("b [ 1]", "d")))
@@ -114,11 +117,11 @@ test_that("set_pars", {
 
 test_that("set_pars missing values", {
   expect_error(set_pars(as.term(c("a [ 1]", "b")), c("b", NA)),
-    "^`value` must not have any missing values[.]$",
-    class = "chk_error"
+               "^`value` must not have any missing values[.]$",
+               class = "chk_error"
   )
   expect_error(set_pars(NA_term_, "a"), "^`x` must not have any missing values[.]$",
-    class = "chk_error"
+               class = "chk_error"
   )
   expect_error(
     set_pars(as.term(c("c c", "b")), "a"),
@@ -132,7 +135,7 @@ test_that("set_pars missing values", {
   expect_identical(
     term,
     structure(c("c [1]", "c[3,2]", "d", "cd"),
-      class = c("term", "character")
+              class = c("term", "character")
     )
   )
 })