+++ NEWS.md
@@ -2,8 +2,8 @@
 - Soft deprecated `is.incomplete_terms()` for `is_incomplete_terms()`.
 - Soft deprecated `is.term()` for `is_term()`.
 - Soft deprecated 'class' value of `validate` argument of `chk_term()` and `vld_term()` and set to be 'complete'.
-- Added `scalar_term()` to test if each term element is scalar.
-- Added `pars_terms()`.
+- `Added `scalar_term()` to test if each term element is scalar.
+- Soft deprecated `pars(x, terms = TRUE)` for `pars_terms(x)`.
 - Migrated `ndims()` to extras package.
 - Migrated `dims()` to extras package.